Additional assertions for SLM restart tests (#48428)
Adds the ability to conditionally wipe SLM policies
and ensures that you can read what you wrote across
restarts.

+++ test/framework/src/main/java/org/elasticsearch/test/rest/ESRestTestCase.java
@@ -473,10 +473,19 @@ public abstract class ESRestTestCase extends ESTestCase {
 
     /**
      * Returns whether to preserve ILM Policies of this test. Defaults to not
-     * preserviing them. Only runs at all if xpack is installed on the cluster
+     * preserving them. Only runs at all if xpack is installed on the cluster
      * being tested.
      */
     protected boolean preserveILMPoliciesUponCompletion() {
+        return false;
+    }
+
+    /**
+     * Returns whether to preserve SLM Policies of this test. Defaults to not
+     * preserving them. Only runs at all if xpack is installed on the cluster
+     * being tested.
+     */
+    protected boolean preserveSLMPoliciesUponCompletion() {
         return false;
     }
 
@@ -514,7 +523,10 @@ public abstract class ESRestTestCase extends ESTestCase {
 
         // Clean up SLM policies before trying to wipe snapshots so that no new ones get started by SLM after wiping
         if (nodeVersions.first().onOrAfter(Version.V_7_4_0)) { // SLM was introduced in version 7.4
-            deleteAllSLMPolicies();
+            if (preserveSLMPoliciesUponCompletion() == false) {
+                // Clean up SLM policies before trying to wipe snapshots so that no new ones get started by SLM after wiping
+                deleteAllSLMPolicies();
+            }
         }
 
         SetOnce<Map<String, List<Map<?,?>>>> inProgressSnapshots = new SetOnce<>();
